+++ data/macros/open-monster-importer.js
@@ -1,0 +1,6 @@
+/**
+ ****************************************************************
+ * This macro can be used to open the monster importer
+ ***************************************************************/
+
+new shadowdark.apps.MonsterImporterSD().render(true);