+++ system/src/tours/ShadowdarkTourTheLostCitadel.mjs
@@ -21,10 +21,24 @@ export class ShadowdarkTheLostCitadelTour extends ShadowdarkTour {
 					action: "scrollTo",
 				},
 				{
-					id: "sd-tlc-compendium",
-					selector: "li[data-pack='shadowdark.the-lost-citadel-adventure']",
+					id: "sd-tlc-compendium-folder-1",
+					selector: "li[data-folder-id='xz3RYT3OvgUDGXad'] header",
+					title: "Shadowdark System folder",
+					content: "<p>Open the Shadowdark System folder.</p>",
+					action: "click",
+				},
+				{
+					id: "sd-tlc-compendium-folder-2",
+					selector: "li[data-folder-id='zzJ6dRCPvUhG0A1V'] header",
+					title: "Quickstart folder",
+					content: "<p>Open the Quickstart folder.</p>",
+					action: "click",
+				},
+				{
+					id: "sd-tlc-compendium-item",
+					selector: "li[data-pack='shadowdark.quickstart-adventures']",
 					title: "Adventure Compendium",
-					content: "<p>The interesting today is the Adventure Compendiums, specifically the Lost Citadel</p>",
+					content: "<p>Today we're interested in the Quickstart Adventure Compendium.</p>",
 					action: "click",
 				},
 				{
@@ -53,7 +67,7 @@ export class ShadowdarkTheLostCitadelTour extends ShadowdarkTour {
 					selector: "#tours-management .window-title",
 					title: "Thank you!",
 					content:
-            "<p><b>Thank you!</b> for following along, learning how to <b>roll from Player character sheet</b> for <b>Shadowdark RPG</b>.</p>\n<p>For more information, see the other available tours.</p>",
+            "<p><b>Thank you!</b> for following along, learning how to <b>import the Quickstart adventure into Foundry</b> for <b>Shadowdark RPG</b>.</p>\n<p>For more information, see the other available tours.</p>",
 					action: "scrollTo",
 				},
 			],